API-5: make sure that only the owner can delete

+++ app/Http/Controllers/Question/DeleteController.php
@@ -7,9 +7,9 @@ use App\Models\Question;
 
 class DeleteController extends Controller
 {
-
     public function __invoke(Question $question)
     {
+        $this->authorize('forceDelete', $question);
         $question->forceDelete();
 
         return response()->noContent();
